Update version to 2. (#84)
