+++ gradle.properties
@@ -1,5 +1,5 @@
 GROUP=com.lyft
-VERSION_NAME=1.1.2-SNAPSHOT
+VERSION_NAME=2.0.0-SNAPSHOT
 
 POM_DESCRIPTION=Android image cropping library.
