+++ scissors-sample/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-    <com.lyft.android.scissors.CropView
+    <com.lyft.android.scissors2.CropView
         android:id="@+id/crop_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
