--- scissors/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=scissors
-POM_NAME=Scissors Library
-POM_PACKAGING=aar
