--- scissors/src/main/java/com/lyft/android/scissors/BitmapLoader.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (C) 2015 Lyft, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.lyft.android.scissors;
-
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.widget.ImageView;
-
-/**
- * Load extension delegates actual Bitmap loading to a BitmapLoader allowing it to use different implementations.
- *
- * @see PicassoBitmapLoader
- * @see GlideBitmapLoader
- */
-public interface BitmapLoader {
-
-    void load(@Nullable Object model, @NonNull ImageView view);
-}
