+++ scissors2/build.gradle
