+++ scissors2/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=scissors2
+POM_NAME=Scissors Library
+POM_PACKAGING=aar
