+++ scissors2/src/main/AndroidManifest.xml
@@ -1,0 +1,3 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<manifest package="com.lyft.android.scissors2" />
