+++ scissors2/src/main/res/values/attrs.xml
