+++ settings.gradle
@@ -1,1 +1,1 @@
-include 'scissors', 'scissors-sample'
+include 'scissors2', 'scissors-sample'
